SPUI-664 > Formatting of prices and amounts (#387)
* Add base formatting functions

* Add decimal to models

* update to flutter 2.10

* add todo

+++ lib/app/shared/features/convert/notifier/preview_buy_with_asset_notifier/preview_convert_notifier.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:logging/logging.dart';
@@ -41,7 +42,7 @@ class PreviewConvertNotifier extends StateNotifier<PreviewConvertState> {
 
   void _updateFrom(PreviewConvertInput input) {
     state = state.copyWith(
-      fromAssetAmount: double.parse(input.amount),
+      fromAssetAmount: Decimal.parse(input.amount),
       fromAssetSymbol: input.fromCurrency.symbol,
       toAssetSymbol: input.toCurrency.symbol,
     );
