+++ lib/app/shared/features/send_by_phone/notifier/send_by_phone_preview_notifier/send_by_phone_preview_notifier.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:logging/logging.dart';
@@ -54,7 +55,7 @@ class SendByPhonePreviewNotifier
       final model = TransferByPhoneRequestModel(
         requestId: DateTime.now().microsecondsSinceEpoch.toString(),
         assetSymbol: currency.symbol,
-        amount: double.parse(state.amount),
+        amount: Decimal.parse(state.amount),
         toPhoneBody: number.body,
         toPhoneCode: number.dialCode,
         toPhoneIso: number.isoCode,
